# specs/002-react-admin-frontend/contracts/openapi-admin.yaml
openapi: 3.0.3
info:
  title: Admin API Endpoints
  description: Backend API endpoints for administrative operations in multi-tenant system. Frontend-agnostic APIs supporting any client framework.
  version: 1.0.0
  contact:
    name: Backend API Team
  license:
    name: MIT
servers:
  - url: /api/v1
    description: API v1 base path (admin endpoints)
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Tenant:
      type: object
      required: [tenant_id, name, is_active, created_at, updated_at, created_by, updated_by]
      properties:
        tenant_id:
          type: string
          format: uuid
          description: Unique tenant identifier
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Organization name
        is_active:
          type: boolean
          description: Tenant activation status
        settings:
          type: object
          properties:
            max_users:
              type: integer
              minimum: 1
              nullable: true
            features:
              type: array
              items:
                type: string
            branding:
              type: object
              properties:
                logo_url:
                  type: string
                  format: uri
                primary_color:
                  type: string
                  pattern: '^#[0-9A-Fa-f]{6}$'
                secondary_color:
                  type: string
                  pattern: '^#[0-9A-Fa-f]{6}$'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
          format: uuid
        updated_by:
          type: string
          format: uuid
    User:
      type: object
      required: [user_id, tenant_id, email, is_disabled, failed_login_attempts, roles, created_at, updated_at, created_by, updated_by]
      properties:
        user_id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          description: Unique within tenant
        is_disabled:
          type: boolean
        failed_login_attempts:
          type: integer
          minimum: 0
        last_login_at:
          type: string
          format: date-time
          nullable: true
        full_name:
          type: string
          nullable: true
        job_title:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        timezone:
          type: string
          description: IANA timezone identifier
          nullable: true
        language:
          type: string
          description: ISO 639-1 language code
          nullable: true
        roles:
          type: array
          items:
            type: string
            enum: [superadmin, tenant_admin, developer, analyst, user, service_account, support_readonly]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
          format: uuid
        updated_by:
          type: string
          format: uuid
...
    ListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items: {}
        pagination:
          type: object
          required: [page, per_page, total, total_pages]
          properties:
            page:
              type: integer
              minimum: 1
            per_page:
              type: integer
              minimum: 1
              maximum: 100
            total:
              type: integer
              minimum: 0
            total_pages:
              type: integer
              minimum: 0
    SingleResponse:
      type: object
      required: [data]
      properties:
        data: {}
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message]
          properties:
            message:
              type: string
            code:
              type: string
            details:
              type: object

    # Create/Update Request Schemas
    CreateTenantRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        is_active:
          type: boolean
          default: true
        settings:
          type: object
    UpdateTenantRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        is_active:
          type: boolean
        settings:
          type: object
    CreateUserRequest:
      type: object
      required: [email, roles]
      properties:
        email:
          type: string
          format: email
        full_name:
          type: string
        job_title:
          type: string
        department:
          type: string
        phone:
          type: string
        timezone:
          type: string
        language:
          type: string
        roles:
          type: array
          items:
            type: string
        is_disabled:
          type: boolean
          default: false
        password:
          type: string
          minLength: 8
          description: Optional - if not provided, invitation will be sent
...
  parameters:
    TenantIdQuery:
      name: tenant_id
      in: query
      description: Filter by tenant ID (required for superadmin cross-tenant operations)
      schema:
        type: string
        format: uuid
    PageQuery:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageQuery:
      name: per_page
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    SortQuery:
      name: sort
      in: query
      description: Sort field
      schema:
        type: string
        default: created_at
    OrderQuery:
      name: order
      in: query
      description: Sort direction
      schema:
        type: string
        enum: [asc, desc]
        default: desc
paths:
...
  /tenants:
    get:
      summary: List tenants
      description: List all tenants (superadmin) or current tenant (tenant_admin)
      parameters:
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/PerPageQuery'
        - $ref: '#/components/parameters/SortQuery'
        - $ref: '#/components/parameters/OrderQuery'
      responses:
        '200':
          description: List of tenants
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ListResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Tenant'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create tenant
      description: Create a new tenant (superadmin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: Tenant created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Tenant'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions (non-superadmin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /tenants/{tenant_id}:
    parameters:
      - name: tenant_id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      summary: Get tenant by ID
      description: Get specific tenant details
      responses:
        '200':
          description: Tenant details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Tenant'
        '404':
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update tenant
      description: Update tenant details (superadmin or tenant_admin for own tenant)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTenantRequest'
      responses:
        '200':
          description: Tenant updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Tenant'

    delete:
      summary: Delete tenant
      description: Delete tenant (superadmin only)
      responses:
        '204':
          description: Tenant deleted
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # User Management Endpoints
  /users:
    get:
      summary: List users
      description: List users within authorized scope (tenant-filtered for non-superadmin)
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/PerPageQuery'
        - $ref: '#/components/parameters/SortQuery'
        - $ref: '#/components/parameters/OrderQuery'
        - name: role
          in: query
          description: Filter by role
          schema:
            type: string
        - name: is_disabled
          in: query
          description: Filter by status
          schema:
            type: boolean
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ListResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'

    post:
      summary: Create user
      description: Create a new user within authorized tenant scope
      parameters:
        - $ref: '#/components/parameters/TenantIdQuery'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
...